Add options to avoid actual bucket deletion (#241)

+++ plugins/public/cinq-auditor-required-tags/cinq_auditor_required_tags/providers.py
@@ -172,7 +172,8 @@ def delete_s3_bucket(client, resource):
         `ActionStatus`
     """
 
-    client.delete_bucket(Bucket=resource.id)
+    if dbconfig.get('enable_delete_s3_buckets', NS_AUDITOR_REQUIRED_TAGS, False):
+        client.delete_bucket(Bucket=resource.id)
     return ActionStatus.SUCCEED, resource.metrics()
